+++ phoenix-server/src/main/java/com/gitee/pifeng/monitoring/server/business/server/entity/MonitorNetHistory.java
@@ -66,7 +66,7 @@ public class MonitorNetHistory {
      * 平均响应时间（毫秒）
      */
     @TableField("AVG_TIME")
-    private Long avgTime;
+    private Double avgTime;
 
     /**
      * 离线次数
