+++ phoenix-ui/src/main/java/com/gitee/pifeng/monitoring/ui/business/web/entity/MonitorNetHistory.java
@@ -56,7 +56,7 @@ public class MonitorNetHistory implements Serializable {
 
     @ApiModelProperty(value = "平均响应时间（毫秒）")
     @TableField("AVG_TIME")
-    private Long avgTime;
+    private Double avgTime;
 
     @ApiModelProperty(value = "离线次数")
     @TableField("OFFLINE_COUNT")
