+++ phoenix-ui/src/main/java/com/gitee/pifeng/monitoring/ui/business/web/vo/MonitorNetVo.java
@@ -54,7 +54,7 @@ public class MonitorNetVo implements ISuperBean {
     private String status;
 
     @ApiModelProperty(value = "平均响应时间（毫秒）")
-    private Long avgTime;
+    private Double avgTime;
 
     @ApiModelProperty(value = "离线次数")
     private Integer offlineCount;
